+++ CHANGELOG.md
@@ -11,6 +11,16 @@ This changelog follows the [Keep a Changelog](http://keepachangelog.com) format.
 
 ## [Unreleased]
 
+### Added ###
+- [#28] Added code `Invalid` (-1).
+- [#28]{Qt C++11} Qt C++11 variant which uses `enum class` for the codes.
+
+### Changed ###
+- {Qt} `networkErrorToStatusCode()` now returns a `Code` instead of `int`.
+
+### Removed ###
+- Testing with Visual Studio 2013 because of outdated CMake.
+
 
 ---